Refactor timer. Create Focus and Break components to be condiutionally rendered in one ternary.#
